Little Big Adventure 2 — Update authors list + Fix archive version number + Improve separation of files between binaries and data packages + Update script to 2.10 syntax

+++ play.it-2/games/play-little-big-adventure-2.sh
@@ -3,6 +3,8 @@ set -o errexit
 
 ###
 # Copyright (c) 2015-2018, Antoine Le Gonidec
+# Copyright (c) 2018, Solène Huault
+# Copyright (c) 2018, Andrey Butirsky
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,42 +32,39 @@ set -o errexit
 
 ###
 # Little Big Adventure 2
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20181103.1
+script_version=20181103.2
 
 # Set game-specific variables
 
 GAME_ID='little-big-adventure-2'
 GAME_NAME='Little Big Adventure 2'
 
-ARCHIVES_LIST='ARCHIVE_GOG'
-
 ARCHIVE_GOG='setup_lba2_2.1.0.8.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/little_big_adventure_2'
 ARCHIVE_GOG_MD5='9909163b7285bd37417f6d3c1ccfa3ee'
 ARCHIVE_GOG_SIZE='750000'
-ARCHIVE_GOG_VERSION='1.0-gog12.1.0.8'
-
-CONFIG_FILES='./lba2.cfg'
+ARCHIVE_GOG_VERSION='1.0-gog2.1.0.8'
 
 ARCHIVE_DOC_DATA_PATH='app'
-ARCHIVE_DOC_DATA_FILES='./*.pdf ./*.txt'
+ARCHIVE_DOC_DATA_FILES='*.pdf *.txt'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./dos4gw.exe ./language.exe ./install.ini ./lba2.exe ./setup.exe'
+ARCHIVE_GAME_BIN_FILES='*.bat *.cfg *.dos *.exe *.ini drivers'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./*.hqr ./*.ile ./*.obl ./drivers ./lba2.cfg ./lba2.dat ./lba2.dos ./lba2.gog ./lba2.ogg ./save ./setsound.bat ./vox'
+ARCHIVE_GAME_DATA_FILES='*.hqr *.ile *.obl lba2.dat lba2.gog lba2.ogg save vox'
 
 GAME_IMAGE='lba2.dat'
+
+CONFIG_FILES='./*.cfg'
 
 APP_MAIN_TYPE='dosbox'
 APP_MAIN_EXE='lba2.exe'
 APP_MAIN_ICON='lba2.exe'
-APP_MAIN_ICON_RES='16 32 48'
 
 PACKAGES_LIST='PKG_BIN PKG_DATA'
 
@@ -77,38 +76,48 @@ PKG_BIN_DEPS="$PKG_DATA_ID dosbox"
 
 # Load common functions
 
-target_version='2.5'
+target_version='2.10'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: ${XDG_DATA_HOME:="$HOME/.local/share"}
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
+fi
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
 fi
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-# Keep Voices on HD
-sed --in-place 's/\(FlagKeepVoice:\) OFF/\1 ON/' "$PLAYIT_WORKDIR/gamedata/app/lba2.cfg"
-
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"    "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}"   "$PATH_GAME"
-done
+# Extract icons
 
 PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+# Keep Voices on HD
+
+file="${PKG_BIN_PATH}${PATH_GAME}/lba2.cfg"
+pattern='s/\(FlagKeepVoice:\) OFF/\1 ON/'
+sed --in-place "$pattern" "$file"
 
 # Write launchers
